move huge text stories into test stories
